# docs/swagger.yaml
basePath: /
definitions:
...
  model.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  model.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - name
    type: object
...
  model.RoleInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.Role'
        description: 关联
      role_id:
        type: integer
      status:
        description: 1-正常，0-禁用
        type: integer
      updated_at:
        type: string
      username:
        type: string
    required:
    - email
    - real_name
    - role_id
    - username
    type: object
...
  model.UserCreateRequest:
    properties:
      department:
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      real_name:
        type: string
      role_id:
        type: integer
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - real_name
    - role_id
    - username
    type: object
  model.UserInfo:
    properties:
      avatar:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        $ref: '#/definitions/model.RoleInfo'
      status:
        type: integer
      username:
        type: string
    type: object
  model.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.UserLoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/model.UserInfo'
    type: object
...
  model.UserUpdateRequest:
    properties:
      department:
        type: string
      email:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role_id:
        type: integer
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vulnark.com
    name: VulnArk Team
    url: https://vulnark.com
  description: VulnArk漏洞管理系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vulnark.com/terms
  title: VulnArk API
  version: 1.0.0
paths:
...
  /api/v1/change-password:
    post:
      consumes:
      - application/json
      description: 修改当前登录用户的密码
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户管理
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户登录
      tags:
      - 用户管理
...
  /api/v1/profile:
    get:
      description: 获取当前登录用户的资料
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户资料
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的资料
      parameters:
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户资料
      tags:
      - 用户管理
...
  /api/v1/users:
    get:
      description: 获取用户列表（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 用户状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 管理员创建新用户
      parameters:
      - description: 创建用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /api/v1/users/{id}:
    delete:
      description: 管理员删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      description: 根据ID获取用户详情
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 管理员更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"